feat: use new header for `main` app
