+++ apps/lend/src/pages/_app.tsx
@@ -10,6 +10,7 @@ import { HashRouter } from 'react-router-dom'
 import type { AppProps } from 'next/app'
 import { connectWalletLocales, initOnboard } from '@/common/features/connect-wallet'
 import { persister, queryClient } from '@/shared/api/query-client'
+import { ThemeProvider } from 'curve-ui-kit/src/shared/ui/ThemeProvider'
 import GlobalStyle from '@/globalStyle'
 import Page from '@/layout/index'
 import { dynamicActivate, initTranslation } from '@/lib/i18n'
@@ -21,7 +22,6 @@ import { QueryProvider } from '@/ui/QueryProvider'
 import { isMobile, removeExtraSpaces } from '@/utils/helpers'
 import { getLocaleFromUrl } from '@/utils/utilsRouter'
 import { getStorageValue } from '@/utils/utilsStorage'
-import { ThemeProvider } from 'curve-ui-kit/src/shared/ui/ThemeProvider'
 
 i18n.load({ en: messagesEn })
 i18n.activate('en')
@@ -62,7 +62,7 @@ function CurveApp({ Component }: AppProps) {
 
     // init theme
     const darkModeQuery = window.matchMedia('(prefers-color-scheme: dark)')
-    updateGlobalStoreByKey('themeType', themeType ? themeType : darkModeQuery.matches ? 'dark' : 'default')
+    updateGlobalStoreByKey('themeType', themeType ? themeType : darkModeQuery.matches ? 'dark' : 'light')
 
     // init locale
     const { rLocale } = getLocaleFromUrl()
@@ -102,8 +102,8 @@ function CurveApp({ Component }: AppProps) {
 
   return (
     <div suppressHydrationWarning>
-      {typeof window === 'undefined' || !appLoaded ? null : (
-        <ThemeProvider theme={themeType as string === 'default' ? 'light' : themeType}>
+      <ThemeProvider theme={themeType as string === 'default' ? 'light' : themeType}>
+        {typeof window === 'undefined' || !appLoaded ? null : (
           <HashRouter>
             <I18nProvider i18n={i18n}>
               <QueryProvider persister={persister} queryClient={queryClient}>
@@ -116,8 +116,8 @@ function CurveApp({ Component }: AppProps) {
               </QueryProvider>
             </I18nProvider>
           </HashRouter>
-        </ThemeProvider>
-      )}
+        )}
+      </ThemeProvider>
     </div>
   )
 }
