+++ apps/main/src/layout/default/index.tsx
@@ -10,25 +10,33 @@ import useLayoutHeight from '@/hooks/useLayoutHeight'
 import useStore from '@/store/useStore'
 
 import Header from '@/layout/default/Header'
-import Footer from '@/layout/default/Footer'
 import GlobalBanner from '@/ui/Banner'
+import { Locale } from '@/ui/AppNav/types'
+import { t } from '@lingui/macro'
+import networks from 'lend/src/networks'
+import Footer from '@/ui/Footer'
 
 const BaseLayout = ({ children }: { children: React.ReactNode }) => {
   const [{ wallet }] = useConnectWallet()
   const globalAlertRef = useRef<HTMLDivElement>(null)
+  const footerRef = useRef<HTMLDivElement>(null)
   useLayoutHeight(globalAlertRef, 'globalAlert')
+  useLayoutHeight(footerRef, 'footer')
 
   const connectState = useStore((state) => state.connectState)
   const isMdUp = useStore((state) => state.isMdUp)
   const layoutHeight = useStore((state) => state.layoutHeight)
   const updateConnectState = useStore((state) => state.updateConnectState)
+  const locale = useStore((state) => state.locale)
+
+  const { rChainId, rNetwork } = getNetworkFromUrl()
+
+  const sections = useMemo(() => getSections(rChainId, locale), [rChainId, locale])
 
   // Update `NEXT_PUBLIC_MAINTENANCE_MESSAGE` environment variable value to display a global message in app.
   const maintenanceMessage = process.env.NEXT_PUBLIC_MAINTENANCE_MESSAGE
 
   const [networkSwitch, setNetworkSwitch] = useState('')
-
-  const { rChainId, rNetwork } = getNetworkFromUrl()
 
   const showSwitchNetworkMessage =
     isFailure(connectState, CONNECT_STAGE.SWITCH_NETWORK) || (!!networkSwitch && isLoading(connectState, networkSwitch))
@@ -60,13 +68,37 @@ const BaseLayout = ({ children }: { children: React.ReactNode }) => {
         handleNetworkChange={handleNetworkChange}
       />
       <Container className={isMdUp ? 'hasFooter' : ''} globalAlertHeight={layoutHeight?.globalAlert}>
-        <Header />
+        <Header sections={sections} />
         <Main minHeight={minHeight}>{children}</Main>
-        {isMdUp && <Footer chainId={rChainId} />}
+        {isMdUp && <Footer sections={sections} ref={footerRef} />}
       </Container>
     </>
   )
 }
+
+const getSections = (rChainId: ChainId, locale: Locale) => [
+  {
+    title: t`Documentation`,
+    links: [
+      { route: 'https://news.curve.fi/', label: t`News` },
+      { route: 'https://resources.curve.fi/lending/understanding-lending/', label: t`User Resources` },
+      { route: 'https://docs.curve.fi', label: t`Developer Resources` },
+      { route: 'https://docs.curve.fi/integration/overview/', label: t`Integrations` },
+      { route: 'https://resources.curve.fi/glossary-branding/branding/', label: t`Branding` },
+      ...(locale === 'zh-Hans' || locale === 'zh-Hant') ? [{ route: 'https://www.curve.wiki/', label: t`Wiki` }] : [],
+    ]
+  },
+  {
+    title: t`Security`, // audits, bug bounty, dune analytics, curve monitor & crvhub
+    links: [
+      { route: 'https://docs.curve.fi/references/audits/', label: t`Audits` },
+      { route: `${networks[rChainId ?? '1']?.orgUIPath}/bugbounty`, label: t`Bug Bounty` },
+      { route: 'https://dune.com/mrblock_buidl/Curve.fi', label: t`Dune Analytics` },
+      { route: 'https://curvemonitor.com', label: t`Curve Monitor` },
+      { route: 'https://crvhub.com/', label: t`Crvhub` }
+    ]
+  }
+]
 
 type MainProps = {
   minHeight: number
