+++ packages/ui/src/index.tsx
@@ -18,6 +18,7 @@ export * from './DetailInfo'
 export * from './Dialog'
 export * from './DialogComboBox'
 export * from './DividerHorizontal'
+export * from './Footer'
 export * from './FormErrorsDisplay'
 export * from './HeaderStats'
 export * from './Icon'
